# 5% faster fs.readdirSync for small directories on macOS 