### What does this PR do?

Read the EOF flag from `getdirentries64`

Yields about a 5% performance improvement when reading small directories by skipping the extra system call to getdirentries since it already told us that there are no more dir entries.

See also:
- https://github.com/apple-oss-distributions/xnu/blob/94d3b452840153a99b38a3a9659680b2a006908e/bsd/vfs/vfs_syscalls.c#L10444-L10470
- https://github.com/apple-oss-distributions/Libc/blob/main/gen/FreeBSD/opendir.c#L373-L392


New:
```zig
cpu: Apple M3 Max
runtime: bun 1.1.17 (arm64-darwin)

benchmark                                              time (avg)             (min … max)       p75       p99      p995
----------------------------------------------------------------------------------------- -----------------------------
readdirSync("/tmp/small-dir", {recursive: false})    7.41 µs/iter     (7.29 µs … 7.62 µs)    7.4 µs   7.62 µs   7.62 µs

 1 files/dirs in /tmp/small-dir
 SHA256: 7ae45ad102eab3b6d7e7896acd08c427a9b25b346470d7bc6507b6481575d519
```

Bun v1.1.16:
```zig
runtime: bun 1.1.16 (arm64-darwin)

benchmark                                              time (avg)             (min … max)       p75       p99      p995
----------------------------------------------------------------------------------------- -----------------------------
readdirSync("/tmp/small-dir", {recursive: false})    7.92 µs/iter     (7.76 µs … 8.03 µs)   7.95 µs   8.03 µs   8.03 µs

 1 files/dirs in /tmp/small-dir
 SHA256: 7ae45ad102eab3b6d7e7896acd08c427a9b25b346470d7bc6507b6481575d519
```

Node.js:
```zig
runtime: node v22.3.0 (arm64-darwin)

benchmark                                              time (avg)             (min … max)       p75       p99      p995
----------------------------------------------------------------------------------------- -----------------------------
readdirSync("/tmp/small-dir", {recursive: false})    7.96 µs/iter     (7.89 µs … 8.24 µs)   7.95 µs   8.24 µs   8.24 µs

 1 files/dirs in /tmp/small-dir
 SHA256: 7ae45ad102eab3b6d7e7896acd08c427a9b25b346470d7bc6507b6481575d519
```

### How did you verify your code works?

There are existing tests, but I also stepped through with the debugger on both large and small directories and verified the received_eof flag is set when it should and when it shouldn't be.